Added a DecayBuffs method, to handle buff durations when turn ends.
Minor tweaks to layout and general structure of the program were done as
well as a method to handle things that happen when turns end.
Also added a test buff in the Program class to test if duration were
decremented as expected - we can further use this to test other things
with buffs such as their effect.